requirements.txt
pandas>=2.2,<3.0
numpy>=1.26,<3.0
PyYAML>=6.0,<7.0
matplotlib>=3.8,<4.0
seaborn>=0.13,<1.0
scipy>=1.12,<2.0
jinja2>=3.1,<4.0
fastapi>=0.110,<1.0
uvicorn>=0.29,<1.0
python-multipart>=0.0.9,<1.0
pytest>=8.0,<9.0
pytest-cov>=5.0,<6.0
chardet>=5.2,<6.0
loguru>=0.7,<1.0
missingno>=0.5,<1.0
streamlit>=1.37,<2.0
